objects/teeth_trap: merge teeth_trap control in TRX
