+++ src/libtrx/meson.build
@@ -257,6 +257,7 @@ sources = [
   'game/objects/traps/pendulum.c',
   'game/objects/traps/rolling_ball.c',
   'game/objects/traps/spikes.c',
+  'game/objects/traps/teeth_trap.c',
   'game/objects/vars.c',
   'game/option/common.c',
   'game/option/controls.c',
